Split `constants.ts` to each files
