+++ pkg/app/web/src/constants/common.ts
@@ -1,0 +1,1 @@
+export const APP_NAME = "PipeCD";
