+++ pkg/app/web/src/constants/path.ts
@@ -1,6 +1,3 @@
-// TODO: move constants directory
-export const APP_NAME = "PipeCD";
-
 export const PAGE_PATH_TOP = "/";
 export const PAGE_PATH_DEPLOYMENTS = "/deployments";
 export const PAGE_PATH_APPLICATIONS = "/applications";
